requirements.txt
docx-mailmerge==0.4.0
lxml==4.2.4
Flask==0.12.2
